Finished progress feature
I finally finished the progress feature. In the habit class there are
three functions: markDone, markUndone and isMarked. Each of them take a
date object as input. markDone and markUndone add/remove the given date
to the habit's progress array. isMarked checks if the given date is in
the habit's progress array.